Convert tulA-kAvErI-snAna-ArambhaH to TOML (month-transition + kaala); move heavy per-conversion tests out of the CI suite
tulA-kAvErI-snAna-ArambhaH: second festival driven by
apply_month_transition_events (same mechanism as muDavan2_muzhukku,
kaala="braahma" instead of "sunrise"). Removed
SolarFestivalAssigner.assign_month_day_tulA_kAvErI_snAna_ArambhaH.

Fixed a real bug found while converting: a bare `return` right after
the original day-loop's first match exited the whole function, so
across a multi-year Panchaanga it only ever assigned ONE occurrence
total, never recurring annually as the festival clearly should.
User-confirmed and fixed as part of this conversion; verified the fix
actually recurs (>=15 occurrences over a 20-year range) rather than
just silently preserving the bug.

Also reorganizes the test suite: the 4 heavy per-conversion
verification tests added this session (anga_intersection_test.py,
intersection_conversion_test.py, month_transition_test.py,
vaara_conditioned_test.py) each build one or more full multi-year
Panchaanga objects from scratch (7-11+ minutes apiece), which pushed
CI runs past an hour. Moved to a new jyotisha_manual_tests/ directory
that `pytest jyotisha_tests` (what CI runs) does not sweep -- kept for
occasional manual validation (`pytest jyotisha_manual_tests`) rather
than every push. The regular CI safety net remains
jyotisha_tests/spatio_temporal (full-panchaanga golden-fixture
comparisons against generated ICS/TeX/MD/JSON), plus the pre-existing
lighter unit tests.

Regenerated the golden spatio_temporal fixtures to reflect the moved
TOML file's URL and the newly-populated timing summary text.
