+++ jyotisha_manual_tests/temporal/festival/applier/anga_intersection_test.py
