+++ jyotisha_manual_tests/temporal/festival/applier/vaara_conditioned_test.py
